+++ app/src/main/java/ru/art241111/dish_recipes/view/searchDishActivity/searchDishActivity.kt
@@ -9,6 +9,8 @@ import android.widget.ImageView
 import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
+import androidx.fragment.app.FragmentManager
+import androidx.fragment.app.FragmentTransaction
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProviders
 import androidx.recyclerview.widget.LinearLayoutManager
@@ -17,7 +19,9 @@ import ru.art241111.dish_recipes.adapters.dishesRecyclerViewAdapter.DishesRecycl
 import ru.art241111.dish_recipes.adapters.dishesRecyclerViewAdapter.OnDataEnd
 import ru.art241111.dish_recipes.adapters.dishesRecyclerViewAdapter.OnItemClickListener
 import ru.art241111.dish_recipes.databinding.ActivitySearchDishBinding
+import ru.art241111.dish_recipes.view.searchDishActivity.fragments.SearchDishesByIngredientsFragment
 import ru.art241111.dish_recipes.view.viewDishActivity.ViewDishActivity
+import ru.art241111.dish_recipes.view.viewDishActivity.fragments.IngredientsAndRecipeInfoFragment
 import ru.art241111.dish_recipes.view_models.SearchDishViewModel
 
 
@@ -47,98 +51,21 @@ class SearchDishActivity : AppCompatActivity(), OnItemClickListener, OnDataEnd {
         // Customization RecycleView: set layoutManager, adapter, data.
         customizationRecycleView()
 
-        // Set click Listener on add button.
-        onAddButtonClickListener()
-
-        // Set click listener on keyboard enter
-        onEnterKeyboardListener()
-
-        // Ingredients recovery after app death.
-        recoveryIngredients()
+        // Add fragment for search dishes by ingredients
+        addFragmentSearchDishesByIngredients()
     }
 
-    /**
-     * Method to add and search ingredients
-     */
-    private fun addIngredients(){
-        // Add ingredients to array.
-        viewModel.addIngredient(binding.etIngredients.text.toString())
+    private fun addFragmentSearchDishesByIngredients() {
+        // Get instance FragmentTransaction.
+        val fragmentManager: FragmentManager = supportFragmentManager
+        val fragmentTransaction: FragmentTransaction = fragmentManager.beginTransaction()
 
-        // Create ingredient view.
-        addIngredientToFlowLayout(binding.etIngredients.text.toString())
+        // Create fragment.
+        val fragment = SearchDishesByIngredientsFragment.newInstance()
 
-        // Clear EditText.
-        binding.etIngredients.text.clear()
-
-        // Load new data
-        viewModel.loadDishesWhenUserAddNewIngredientOrStartApplication()
-    }
-
-    /**
-     * Set click listener on enter keyboard
-     */
-    private fun onEnterKeyboardListener() {
-        binding.etIngredients.setOnKeyListener(
-                View.OnKeyListener { _, keyCode, event ->
-                    if (keyCode == KeyEvent.KEYCODE_ENTER && event.action == KeyEvent.ACTION_UP) {
-                        addIngredients()
-                        return@OnKeyListener true
-                    }
-                    false
-                }
-        )
-    }
-
-    /**
-     * Click listener of add button.
-     */
-    private fun onAddButtonClickListener() {
-        binding.ibAddIngredients.setOnClickListener{
-            addIngredients()
-        }
-    }
-
-    /**
-     * Ingredients recovery after app death.
-     */
-    private fun recoveryIngredients() {
-        for (ingredient in viewModel.ingredients){
-            addIngredientToFlowLayout(ingredient)
-        }
-    }
-
-    /**
-     * Add ingredient to flow layout.
-     * @param ingredientName - name of ingredient.
-     */
-    private fun addIngredientToFlowLayout(ingredientName: String) {
-        val inflater = layoutInflater
-        val ingredient = inflater.inflate(R.layout.ingredient, null)
-
-        // Customization ingredients view.
-        fillingInDataToIngredientView(ingredient, ingredientName)
-
-        // Add ingredient to FlowLayout.
-        binding.flListIngredients.addView(ingredient)
-    }
-
-    /**
-     * Filling in the view data.
-     * @param ingredient - new view.
-     * @param ingredientName - name of ingredient.
-     */
-    private fun fillingInDataToIngredientView(ingredient: View, ingredientName: String) {
-        // Set ingredients name.
-        val textView = ingredient.findViewById<TextView>(R.id.tv_name_ingredient)
-        textView.text = ingredientName
-
-        // Add click listener om delete button.
-        val imageView = ingredient.findViewById<ImageView>(R.id.btn_delete)
-        imageView.setOnClickListener {
-            binding.flListIngredients.removeView(ingredient)
-            viewModel.deleteIngredient(ingredientName)
-            viewModel.loadDishesWhenUserAddNewIngredientOrStartApplication()
-        }
+        // Add fragment to LinearLayout.
+        fragmentTransaction.add(R.id.ll_main_search_dish_activity, fragment)
+        fragmentTransaction.commit()
     }
 
     /**
