+++ app/src/main/res/layout/fragment_search_dishes_by_ingredients.xml
@@ -1,0 +1,48 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    tools:context=".view.searchDishActivity.fragments.SearchDishesByIngredientsFragment"
+    >
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:orientation="vertical">
+
+        <LinearLayout
+            android:id="@+id/ll_add_ingredients"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_marginBottom="5dp"
+            android:orientation="horizontal">
+
+            <EditText
+                android:id="@+id/et_ingredients"
+                android:layout_width="0dp"
+                android:layout_height="wrap_content"
+                android:hint = "@string/enter_ingredients"
+                android:ems="10"
+                android:layout_weight="10"
+                android:maxLines="1"
+                android:inputType="text"
+                android:autofillHints="" />
+
+            <ImageButton
+                android:id="@+id/ib_add_ingredients"
+                android:layout_width="40dp"
+                android:layout_height="40dp"
+
+                android:src="@drawable/ic_add_white_24dp"
+                android:background="@drawable/background_button"
+                android:contentDescription="@string/add_ingredient" />
+
+        </LinearLayout>
+
+        <org.apmem.tools.layouts.FlowLayout
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+
+            android:id="@+id/fl_list_ingredients" >
+        </org.apmem.tools.layouts.FlowLayout>
+
+    </LinearLayout>
+</layout>
